requirements.txt
openai
numpy
torch==2.2.2
transformers
optimum
python-dotenv
gradio
llama-index-core
llama-index-readers-file
llama-index-embeddings-openai
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
llama-index-vector-stores-azurecosmosmongo
markdown
pymongo